+++ bsn1/CHANGELOG.md
@@ -393,6 +393,7 @@ Nothing is changed.
 ### Added
 
 - Add variant `Anyhow` to enum `Error`
+- Add variant `InvalidDerBooleanContents` to enum `Error`
 - Implement `From<anyhow::Error>` for `Error`
 - Create method `Error::into\_anyhow()`
 - Create method `Error::root\_cause()`